+++ src/scvi/_constants.py
@@ -3,6 +3,7 @@ from typing import NamedTuple
 
 class _REGISTRY_KEYS_NT(NamedTuple):
     X_KEY: str = "X"
+    ATAC_X_KEY: str = "atac"
     BATCH_KEY: str = "batch"
     SAMPLE_KEY: str = "sample"
     LABELS_KEY: str = "labels"